requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
openai>=1.0.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
flake8 